+++ src/HashException.php
@@ -4,5 +4,4 @@ namespace SecurePassword;
 
 class HashException extends \Exception
 {
-    
 }