+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Released Notes
 
+## v1.0.0 - (2021-10-12)
+
+### Added
+
+- Added settings in construct
+
+### Changed
+
+- Changed HashAlgorithm constant to public
+
+### Fixed
+
+- Fixed bugs
+
+-----------------------------------------------------------
+
 ## v0.2.0 - (2021-04-20)
 
 ### Changed